+++ src/main/java/de/kosit/validationtool/impl/Printer.java
@@ -18,4 +18,14 @@ public class Printer {
     public static void writeOut(final String message, final Object... params) {
         System.out.println(MessageFormat.format(message, params));
     }
+
+    /**
+     * Writes to standard error channel.
+     *
+     * @param message the message with placeholders
+     * @param params the params.
+     */
+    public static void writeErr(final String message, final Object... params) {
+        System.err.println(MessageFormat.format(message, params));
+    }
 }